Windows sysconfiggen paths and pe shim exports

+++ test/struct_abi/src/abi_fixture.cpp
@@ -7,7 +7,15 @@
 //   H1 {void*}      8B, one INTEGER eightbyte  → RAX / RDI
 //   P2 {int,int}    8B, two ints share one eightbyte (RAX packs both)
 //   F2 {float,float}8B, one SSE eightbyte      → XMM0 (both floats)
-//   B3 {long x3}   24B, MEMORY class           → sret, and byval as an ARG
+//   B3 {long long x3} 24B, MEMORY class        → sret, and byval as an ARG
+//                  `long long`, NOT `long`: this file states sizes, and `long`
+//                  is the one integer the word size does not settle (4 bytes
+//                  under LLP64, 8 under Unix64). As `long` the struct was 12
+//                  bytes on Windows while the traces below — hand-written MLIR
+//                  at i64, and NTuple{3,Int64} on the Julia side — still said
+//                  24, so b3_make returned (0x8_00000007, …) for (7, 8, 9).
+//                  The shape under test is MEMORY-class-by-sret, not the width
+//                  of a long; test_llp64_widths.jl owns that question.
 //   Gap            24B, MEMORY class WITH interior padding — the shape the
 //                  emitted-size bug mis-modelled (2026-08-05)
 
@@ -24,8 +32,8 @@ typedef struct { float x, y; } F2;
 F2 f2_make(float x, float y) { F2 s; s.x = x; s.y = y; return s; }
 float f2_sum(F2 v) { return v.x + v.y; }
 
-typedef struct { long a, b, c; } B3;
-B3 b3_make(long a, long b, long c) { B3 s; s.a = a; s.b = b; s.c = c; return s; }
+typedef struct { long long a, b, c; } B3;
+B3 b3_make(long long a, long long b, long long c) { B3 s; s.a = a; s.b = b; s.c = c; return s; }
 
 // MEMORY-class struct as an ARGUMENT. SysV wants a caller-owned copy in the
 // outgoing stack argument area (`byval`); passing the aggregate as an LLVM
@@ -33,7 +41,7 @@ B3 b3_make(long a, long b, long c) { B3 s; s.a = a; s.b = b; s.c = c; return s; 
 // and passing a bare pointer hands the callee an address where it expects
 // bytes. Only a system-clang callee can catch either — a self-JIT'd one shares
 // whatever convention the JIT chose.
-long b3_sum(B3 v) { return v.a + v.b + v.c; }
+long long b3_sum(B3 v) { return v.a + v.b + v.c; }
 
 // Interior padding (4 bytes after `a`) plus a trailing bool run, then tail
 // padding: 24 bytes whose member sizes sum to only 19. That gap is what the
@@ -48,8 +56,8 @@ Gap gap_make(int a, void* p, int b) {
 
 // Every field participates, so a mis-marshalled by-value copy is a wrong
 // NUMBER rather than a crash we might read as something else.
-long gap_probe(Gap g) {
-    return (long)g.a * 1000000L + (long)g.b * 1000L
+long long gap_probe(Gap g) {
+    return (long long)g.a * 1000000LL + (long long)g.b * 1000LL
          + (g.f1 ? 1 : 0) + (g.f2 ? 2 : 0) + (g.f3 ? 4 : 0)
          + (g.p == (void*)0x1234 ? 100 : 0);
 }
